feat: Use -d for safer branch deletion
This commit changes the `git branch -D` command to `git branch -d` when deleting the merged feature branch.

This is a safer deletion flag that prevents accidental data loss if a commit is made to the feature branch after it is pushed but before it is deleted. This can happen in a race condition where another developer commits to the branch between our push and the deletion. If this occurs, `-d` will refuse to delete the branch, preventing the loss of that commit.

A new test `test_merge_race_condition_commit_after_push` has been added to cover this scenario.